refactor: change from long double to double because if ABI changes in gcc 4.4

+++ src/rlsl_token.h
@@ -104,7 +104,7 @@ typedef struct rlsl_token_t {
             bool overflow;
         } num_int;
         struct {
-            long double value;
+            double value;
             bool overflow;
         } num_float;
         char* identifier;
